mosquitto_python_and_notion/requirements.txt
notion-client
paho-mqtt